# Conflicts:
#	GRADUATION PROJECT/MyFrameWork/DreamGame/DreamGame.vcxproj.filters
#	GRADUATION PROJECT/MyFrameWork/DreamGame/GameObject.cpp
#	GRADUATION PROJECT/MyFrameWork/DreamGame/GameobjectManager.cpp

+++ GRADUATION PROJECT/MyFrameWork/DreamGame/Camera.cpp
@@ -5,8 +5,8 @@ CCamera::CCamera()
 {
 	m_xmf4x4View = Matrix4x4::Identity();
 	m_xmf4x4Projection = Matrix4x4::Identity();
-	m_d3dViewport = { 0, 0, FRAME_BUFFER_WIDTH , FRAME_BUFFER_HEIGHT, 0.0f, 1.0f };
-	m_d3dScissorRect = { 0, 0, FRAME_BUFFER_WIDTH , FRAME_BUFFER_HEIGHT };
+	m_d3dViewport = { 0, 0, float(FRAME_BUFFER_WIDTH) , float(FRAME_BUFFER_HEIGHT), 0.0f, 1.0f };
+	m_d3dScissorRect = { 0, 0, LONG(FRAME_BUFFER_WIDTH) , LONG(FRAME_BUFFER_HEIGHT) };
 	m_xmf3Position = XMFLOAT3(0.0f, 0.0f, 0.0f);
 	m_xmf3Right = XMFLOAT3(1.0f, 0.0f, 0.0f);
 	m_xmf3Look = XMFLOAT3(0.0f, 0.0f, 1.0f);
@@ -30,8 +30,8 @@ CCamera::CCamera(CCamera* pCamera)
 	{
 		m_xmf4x4View = Matrix4x4::Identity();
 		m_xmf4x4Projection = Matrix4x4::Identity();
-		m_d3dViewport = { 0, 0, FRAME_BUFFER_WIDTH , FRAME_BUFFER_HEIGHT, 0.0f, 1.0f };
-		m_d3dScissorRect = { 0, 0, FRAME_BUFFER_WIDTH , FRAME_BUFFER_HEIGHT };
+		m_d3dViewport = { 0, 0, float(FRAME_BUFFER_WIDTH) , float(FRAME_BUFFER_HEIGHT), 0.0f, 1.0f };
+		m_d3dScissorRect = { 0, 0, LONG(FRAME_BUFFER_WIDTH) , LONG(FRAME_BUFFER_HEIGHT) };
 		m_xmf3Position = XMFLOAT3(0.0f, 0.0f, 0.0f);
 		m_xmf3Right = XMFLOAT3(1.0f, 0.0f, 0.0f);
 		m_xmf3Look = XMFLOAT3(0.0f, 0.0f, 1.0f);
@@ -160,8 +160,8 @@ void CCamera::ReInitCamrea()
 {
 	m_xmf4x4View = Matrix4x4::Identity();
 	m_xmf4x4Projection = Matrix4x4::Identity();
-	m_d3dViewport = { 0, 0, FRAME_BUFFER_WIDTH , FRAME_BUFFER_HEIGHT, 0.0f, 1.0f };
-	m_d3dScissorRect = { 0, 0, FRAME_BUFFER_WIDTH , FRAME_BUFFER_HEIGHT };
+	m_d3dViewport = { 0, 0, float(FRAME_BUFFER_WIDTH) , float(FRAME_BUFFER_HEIGHT), 0.0f, 1.0f };
+	m_d3dScissorRect = { 0, 0, LONG(FRAME_BUFFER_WIDTH) ,LONG( FRAME_BUFFER_HEIGHT) };
 	m_xmf3Position = XMFLOAT3(0.0f, 0.0f, 0.0f);
 	m_xmf3Right = XMFLOAT3(1.0f, 0.0f, 0.0f);
 	m_xmf3Look = XMFLOAT3(0.0f, 0.0f, 1.0f);
